seperating files to keep clean code base and prep for moving async functions --> RTK query
