+++ src/screens/map-screen.tsx
@@ -18,7 +18,7 @@ import { Product } from 'src/cards/product';
 import { BasicProductData } from 'src/types/products';
 import { converters } from 'src/firebase/db-converters';
 
-// import { fetchCloseData } from 'src/firebase/products-api';
+import { fetchCloseData } from 'src/firebase/map-api';
 // import { useFetchLocationProductsQuery } from 'src/services/products-queries';
 
 type Location = [number, number];
@@ -63,89 +63,59 @@ export const MapScreen = () => {
     }, [location]);
 
     // const fetchCloseData = async () => {
-    //     const radius = 4 * 1000;
+    //     const radius = 2000 * 1000;
+    //     // const radius = Infinity;
 
     //     if (location?.length === 2) {
     //         const bounds = geohashQueryBounds(location, radius);
     //         const promises = [];
+
     //         for (const b of bounds) {
-    //             const colRef = collection(db, 'basic-product-data').withConverter<BasicProductData>(
-    //                 converters.productData,
-    //             );
+    //             const colRef = collection(db, 'basic-product-data') as any;
     //             const q = query(colRef, orderBy('geohash'), startAt(b[0]), endAt(b[1]));
-    //             const data = await getDocs(q);
-    //             // console.log(data);
 
-    //             // const prod: BasicProductData[] = [];
+    //             const datas = (await getDocs(q)) as any;
 
-    //             // await Promise.all(
-    //             data.docs.map((doc) => {
-    //                 // console.log(doc.data());
-    //                 const datas = { ...doc.data(), id: doc.id };
-    //                 // prod.push(datas);
-    //                 console.log(datas);
-    //             });
-    //             // );
-
-    //             // return prod;
+    //             promises.push(datas);
     //         }
+
+    //         const thing = await Promise.all(promises)
+    //             .then((snapShots) => {
+    //                 const matchingDocs: any = [];
+
+    //                 for (const snap of snapShots) {
+    //                     for (const doc of snap.docs) {
+    //                         const lat = doc.get('lat');
+    //                         const lng = doc.get('long');
+
+    //                         // console.log(lat);
+    //                         // console.log(lng);
+
+    //                         // We have to filter out a few false positives due to GeoHash
+    //                         // accuracy, but most will match
+    //                         const distanceInKm = distanceBetween([lat, lng], location);
+    //                         const distanceInM = distanceInKm * 1000;
+    //                         if (distanceInM <= radius) {
+    //                             const locationProd = { ...doc.data(), id: doc.id };
+    //                             matchingDocs.push(locationProd);
+    //                         }
+    //                     }
+    //                 }
+    //                 // console.log(matchingDocs);
+    //                 return matchingDocs;
+    //             })
+    //             .then((matchingDocs) => {
+    //                 // console.log(matchingDocs);
+    //                 return matchingDocs;
+    //             });
+    //         // const thingers = await Promise.all(thing)
+    //         // console.log(thing);
+    //         return thing;
     //     }
     // };
 
-    const fetchCloseData = async () => {
-        const radius = 2000 * 1000;
-        // const radius = Infinity;
-
-        if (location?.length === 2) {
-            const bounds = geohashQueryBounds(location, radius);
-            const promises = [];
-
-            for (const b of bounds) {
-                const colRef = collection(db, 'basic-product-data') as any;
-                const q = query(colRef, orderBy('geohash'), startAt(b[0]), endAt(b[1]));
-
-                const datas = (await getDocs(q)) as any;
-
-                promises.push(datas);
-            }
-
-            const thing = await Promise.all(promises)
-                .then((snapShots) => {
-                    const matchingDocs: any = [];
-
-                    for (const snap of snapShots) {
-                        for (const doc of snap.docs) {
-                            const lat = doc.get('lat');
-                            const lng = doc.get('long');
-
-                            // console.log(lat);
-                            // console.log(lng);
-
-                            // We have to filter out a few false positives due to GeoHash
-                            // accuracy, but most will match
-                            const distanceInKm = distanceBetween([lat, lng], location);
-                            const distanceInM = distanceInKm * 1000;
-                            if (distanceInM <= radius) {
-                                const locationProd = { ...doc.data(), id: doc.id };
-                                matchingDocs.push(locationProd);
-                            }
-                        }
-                    }
-                    // console.log(matchingDocs);
-                    return matchingDocs;
-                })
-                .then((matchingDocs) => {
-                    // console.log(matchingDocs);
-                    return matchingDocs;
-                });
-            // const thingers = await Promise.all(thing)
-            // console.log(thing);
-            return thing;
-        }
-    };
-
     const getPost = async () => {
-        const thing: any = await fetchCloseData();
+        const thing: any = await fetchCloseData(location);
         console.log(`thing ${thing}`);
         console.log(thing);
         setProducts(thing);
